+++ main.js
@@ -1,8 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { createStore, combineReducers } from 'redux';
+import { createStore } from 'redux';
 
-import TodoApp from './src/components/TodoApp';
+import TodoApp from './src/components/';
+import todoApp from './src/reducers/';
 
 export const getVisibleTodos = (todos, filter) => {
   switch(filter) {
@@ -21,63 +22,10 @@ export const getVisibleTodos = (todos, filter) => {
   }
 };
 
-const todo = (state, action) => {
-  switch(action.type) {
-    case 'ADD_TODO':
-      return {
-        id: action.id,
-        text: action.text,
-        completed: false
-      }
-    case 'TOGGLE_TODO':
-      if(action.id !== state.id) {
-        return state;
-      }
-      return {
-        ...state,
-        completed: !state.completed
-      };
-    default:
-      return state;
-  };
-};
-
-const todos = (state = [], action) => {
-  switch(action.type) {
-  case 'ADD_TODO':
-    return [
-      ...state,
-      todo(state, action)
-    ];
-  case 'TOGGLE_TODO':
-    return state.map(t => todo(t, action));
-  default:
-    return state;
-  }
-};
-
-const visibilityFilter = (state = 'SHOW_ALL', action) => {
-  switch(action.type) {
-    case 'SET_VISIBILITY_FILTER':
-      return action.filter;
-    default:
-      return state;
-  }
-};
-
-const todoApp = combineReducers({
-  todos,
-  visibilityFilter
-});
-
 export const store = createStore(todoApp);
 
-const render = () => {
-  ReactDOM.render(
-    <TodoApp {...store.getState()} />,
-    document.getElementById('root')
-  );
-};
+ReactDOM.render(
+  <TodoApp />,
+  document.getElementById('root')
+);
 
-store.subscribe(render);
-render();
