+++ src/reducers/index.js
@@ -1,0 +1,11 @@
+import { combineReducers } from 'redux';
+
+import todos from './todos';
+import visibilityFilter from './visibilityFilter';
+
+const todoApp = combineReducers({
+  todos,
+  visibilityFilter
+});
+
+export default todoApp;
