+++ src/reducers/todos.js
@@ -1,0 +1,17 @@
+import todo from './todo';
+
+const todos = (state = [], action) => {
+  switch(action.type) {
+  case 'ADD_TODO':
+    return [
+      ...state,
+      todo(state, action)
+    ];
+  case 'TOGGLE_TODO':
+    return state.map(t => todo(t, action));
+  default:
+    return state;
+  }
+};
+
+export default todos;
