+++ src/reducers/visibilityFilter.js
@@ -1,0 +1,10 @@
+const visibilityFilter = (state = 'SHOW_ALL', action) => {
+  switch(action.type) {
+    case 'SET_VISIBILITY_FILTER':
+      return action.filter;
+    default:
+      return state;
+  }
+};
+
+export default visibilityFilter;
